Graphs started
Structure for the graph is nearly. It is a different take on graphs. Must people from what i have observed makes a Graph class which handles if it should accept parallel edges and loops and/or if it is a simple graph and so on. I am starting an experminent based on the observation that all these requirements are based on the composition of the graph in terms of its edges. On the grounds of this I will move this responsibility to the edge set class and have a Graph class act like an interface with most actions happening through the edge set class.